pyproject.toml
[tool.poetry]
name = "roman-cuts"
version = "0.1.0"
description = ""
authors = ["Jorge Martinez-Palomera <jorgemarpa@ug.uchile.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = ">1.26,<3.0"
matplotlib = "^3.10.1"
pandas = ">1.5,<3.0"
astropy = ">5,<7"
tqdm = "^4.67.1"
asdf-astropy = ">=0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = ">=0.5,<0.12"
ipywidgets = ">=7.6.3,<9.0.0"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.0"
mkdocstrings = {version = "^0.29.0", extras = ["python", "numpy-style"]}
mkdocs-jupyter = "^0.25.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Assumes Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`), a subset of pycodestyle (`E`) and isort (`I`).
select = ["E4", "E7", "E9", "F", "I"]

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
disable_error_code = ["import-untyped", "index"]